+++ wcsupgrade/metadata.rb
@@ -1,8 +1,8 @@
-name             'wcupgrade'
+name             'wcsupgrade'
 maintainer       'TMW'
 maintainer_email 'YOUR_EMAIL'
 license          'All rights reserved'
-description      'Installs/Configures wcupgrade'
+description      'Installs/Configures wcsupgrade'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 depends		 'wcsattr'